# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Hussla API
  version: 1.0.0
  description: Self-hosted job-search tracker. Agents read and write jobs, companies, answers and drafts; the owner approves email.
servers:
  - url: /
security:
  - agentKey: []
  - localSession: []
...
paths:
...
  /api/mail/settings:
    get:
      tags: [mail]
      operationId: getMailSettings
      summary: Owner only. The saved mail setup; `hasSecret` says whether a password or key is stored, never what it is.
      responses:
        "200":
          description: The setup (all empty when none is saved).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailSettings" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    put:
      tags: [mail]
      operationId: saveMailSettings
      summary: >-
        Owner only, with a passkey tap. Saves the mail setup. `secret` is the app password or API key;
        leave it out to keep the stored one (same provider only). It is stored encrypted and never returned.
      parameters:
        - $ref: "#/components/parameters/StepUp"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MailSettingsSave" }
      responses:
        "200":
          description: Saved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailSettings" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/setup/code:
    post:
      tags: [setup]
      operationId: requestSetupCode
      summary: >-
        Prints a new setup code to the server log, replacing the live one (a lost log, or a lost
        passkey). The owner may ask; before there is an owner, so may anyone who could claim with it.
        At most once a minute (429).
      security: []
      responses:
        "200":
          description: Printed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooSoon" }
...
  /api/passkeys/{passkeyId}:
    parameters:
      - name: passkeyId
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [setup]
      operationId: removePasskey
      summary: Owner only, with a passkey tap. Removes a passkey (a lost phone's). The last passkey for an address can't be removed (409).
      parameters:
        - $ref: "#/components/parameters/StepUp"
      responses:
        "200":
          description: Removed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    agentKey:
      type: http
      scheme: bearer
      description: An agent key made on the Settings page. Shows in the activity log as `agent:<name>`.
    localSession:
      type: apiKey
      in: cookie
      name: hussla_session
      description: >-
        The owner's session on the local listener (http://localhost), started by the one-time link
        `hussla open` prints. On the tailnet the owner needs no credential: embedded Tailscale names
        the connection's user (not modeled by OpenAPI). No header ever names the owner.
  parameters:
    StepUp:
      name: X-Hussla-Step-Up
      in: header
      required: true
      description: A one-use token from POST /api/stepup/finish, granted for exactly this method and path.
      schema: { type: string }
...
  responses:
    BadRequest:
      description: The input is invalid; `error` says which field and why.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No valid agent key or owner identity.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: 'Owner only; a non-owner tailnet user; a foreign Origin, or an owner write without one; no passkey tap (code "passkey-required"); or no owner enrolled yet (code "setup-required").'
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such record.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: 'The record exists, is in a state that doesn''t allow this, changed since the caller read it, or an agent tried to change owner-written fields (see `code` and `fields`).'
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    TooSoon:
      description: Asked again too soon; wait a minute.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code:
          type: string
          description: "A stable reason for the refusals a client acts on."
          enum: [owner-field, changed-since-read, transition-not-allowed, setup-required, passkey-required]
        fields:
          type: array
          items: { type: string }
          description: 'With code "owner-field": the fields the owner last wrote, which an agent may not change.'
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    MailSettings:
      type: object
      required: [configured, hasSecret, providerId, host, port, security, username, fromAddress, fromName, region, domain]
      properties:
        configured: { type: boolean }
        hasSecret: { type: boolean, description: A password or key is stored (it is never returned). }
        providerId: { type: string }
        host: { type: string }
        port: { type: integer }
        security: { type: string }
        username: { type: string }
        fromAddress: { type: string }
        fromName: { type: string }
        region: { type: string }
        domain: { type: string }
    MailSettingsSave:
      type: object
      required: [providerId, fromAddress]
      properties:
        providerId: { type: string }
        host: { type: string }
        port: { type: integer }
        security: { type: string, description: "\"starttls\" or \"tls\"; empty for the provider's own." }
        username: { type: string }
        fromAddress: { type: string }
        fromName: { type: string }
        region: { type: string }
        domain: { type: string }
        secret: { type: string, description: The app password or API key; leave out to keep the stored one. }
...